+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 <for new features.>
 
-- add a function that scroll the scroll view contents until the specified position is visible
+- add a function that scrolls the scroll view contents until the specified position is visible
+- add a safely subscript funciton for protocol `Collection`.
 
 ### Fixed
 <for any bug fixes.>